+++ apps/skel/webapp/controller/base.py
@@ -1,0 +1,9 @@
+from chula.www import controller
+
+class Controller(controller.Controller):
+    def render(self, template):
+        """
+        Add your (Mako, Mustache, Cheetah) logic here
+        """
+
+        return template