+++ apps/skel/webapp/controller/home.py
@@ -1,0 +1,8 @@
+from controller import base
+
+class Home(base.Controller):
+    def index(self):
+        return self.render('Hello world')
+
+    def foo(self):
+        return self.render('Foo')